requirements.txt
# Python Dependencies for Predictive Maintenance MLOps Pipeline

# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning Frameworks
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2

# Azure SDK
azure-ai-ml==1.9.0
azure-identity==1.13.0
azureml-core==1.51.0
azure-storage-blob==12.17.0

# Web Framework & API
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0

# Monitoring & Metrics
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
joblib==1.3.1
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.4.1

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2
